Add UI support for the colormode light attribute
If the light color is set as a color temperature, the hue & sat
values for it do not really match up. Therefore set the color of
the color picker to a partially transparent white in that case.

For colors set using XY mode, the bridge calculates the respective
hue & sat values, but not immediately - so it might take a few
more seconds.