Lookup light name from the Hue API if device name is an empty string
The Hue API provides extended information about the lights available,
including a user friendly name. If the device name is set to the
empty string, use an auto generated name based on the class name
(light type) and the id to satisfy the ancestor constructor, and
change it once the Hue API name is received after the first poll.