Implement FIFO queue (with concurrency) for Hue API requests
The Hue bridge can easily get overloaded in a situation with many
Hue lights (due to periodic polling) and/or many requests being
sent at the same time from the color picker or in rules.
A FIFO queue (using simple-promise-queue) is implemented with a
configurable concurrency (default: 2) and a maximum queue length
(4 + # devices).