Introduce global polling of all lights and groups
Instead of polling configured lights and groups individual every
interval, in many cases it's more efficient to simply poll the
status of all lights and groups in one Hue API request. As of
yesterday this is supported by node-hue-api using the lights()
and groups() methods.
Fixes #2

Also changes the way light state handlers are registered; using
explicit callbacks instead of chaining on a Promise.