Properly restore the (full) light state
Only restoring the previous color value is not always sufficient:
The light may have been implicitly turned on by the action, and
due to different color modes, the saved value may not accurately
represent the previous color (e.g. if changing color when color
mode was set using color temperature.) Therefore, save the
complete light state and restore it.

In the device classes a new methods saveLightState() is introduced
that prepares a hue state change for later restoration by
restoreLightState().