Cleanup Hue API light/group state handling
Hue light and light group state and state changes were being
stored and handled in a variety of ways with by now rather
confusing naming. Simplify and clarify this:

The node-hue-api LightState class was also used in the
BaseHueLight class to represent the current state of a light
or group, which it wasn't designed or convenient for. Rely on
on plain objects with properties for this instead.

Clearly separate the concepts of "current state" and
"state change".

* Remove BaseHueLight.lightState property
* Rename all uses of LightState objects to "stateChange"
* Rename _mergeLightState to _mergeStateChange
* Rename BaseHueLight.setLightState to changeHueState
* Rename BaseHueLight.createLightState to prepareStateChange
* Rename hueState variable to hueStateChange
* Replace all remaining uses of LightState._values with .payload()
* Rename light group result .action or .lastAction to .state
* Clarify debug messages for state changes