Make scene name lookup case insensitive and use last listed scene only
It's unclear what the timestamp in the name field of a Hue API scene
actually means, and it doesn't seem to necessarily be in chronological
order. It *appears* (but isn't specified) that newer scenes are
returned after older scenes, so let's just use the last scene returned
with the same name.