Bound number of retries on global polling requests
Although polling will be repeated infinitely anyway, bound the
number of initial polling requests so an error message is reported
if a poll doesn't succeed after the number of attempts set in the
configuration.
Set the initial retry amount to tenfold for the very first poll.
The error handler sets an advisory delay time on the error object,
so hammering can be avoided even after retries are depleted.