Select the latest Hue scene based on the lastupdated value only
The order in which scenes are returned by the Hue bridge does not
seem representative of the order in which they were defined.
Recent API versions have introduced a "lastupdated" date/timestamp,
which is null for scenes that predated it. Use it if it exists,
and ignore the (seemingly) application specific timestamp in the
name completely, other than stripping it off.