Do not (unintentionally) add plugin config 'name' property
Due to a recent change in Pimatic 0.8.103, pimatic-hue-zll started
adding empty name: "" properties to the plugin configuration.

Support in pimatic was added in pimatic/pimatic@8b5cb6d for a
'prepareConfig' method of the plugin class to prepare the plugin
config. pimatic-hue-zll happened to be using that method name
(predating the pimatic change) to prepare the device config.

Move the device config prepareConfig method to a static method
of the device classes, and introduce a new plugin prepareConfig
method to automatically remove erroneously added plugin config
name properties. Fixes #8