Perform automatic discovery of the Hue bridge on startup
Instead of only performing discovery of the bridge on (manually
initiated) device discovery, do this on startup as well if not
hard coded in the configuration. We no longer automatically add it
to the config, so bridges on (very) dynamic IPs are also
supported.

Initialization of devices (e.g. setting up polling) is now
deferred to a separate device init method, until the hue API is
initialized and available.

Issue #11